% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqatools.R
\name{appender.file2}
\alias{appender.file2}
\title{appender.file2}
\usage{
appender.file2(format, console = FALSE, inherit = TRUE,
  datetime.fmt = "\%Y\%m\%dT\%H\%M\%S")
}
\arguments{
\item{format}{the format to use}

\item{console}{log to console?}

\item{inherit}{inherits other logging proprerties}

\item{datetime}{the datetime}

\item{fmt}{the format to use}
}
\description{
logs to both a file and the console
}
\examples{
flog.threshold(INFO)
flog.appender(appender.file2(paste0("logs/mylog-",Sys.Date(),".log"), console=TRUE))
flog.info("Started...")
}
\keyword{log}
