% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqatools.R
\name{buildSQLInClause}
\alias{buildSQLInClause}
\title{buildSQLInClause}
\usage{
buildSQLInClause(myList, quoteChar = "")
}
\arguments{
\item{myList}{the list to build from}

\item{quoteChar}{a character to quote}
}
\description{
builds an IN criteria clause or a SELECT clause from a list
}
\examples{
buildSQLInClause(c(a,b,c),quoteChar="'")
}
\keyword{IN}
\keyword{builder}
\keyword{clause}
\keyword{criteria}
\keyword{select}
