% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqatools.R
\name{checkDependencies}
\alias{checkDependencies}
\title{checkDependencies}
\usage{
checkDependencies()
}
\description{
Checks dependicies on startup
}
\examples{
# typical
rm(list=ls())
library('eqatools')
eqatools::checkDependencies()

########################################################################
#
#  Global Config (Get our globals.yml file and configure any special params here..)
#
globals <- config::get(file = "globals.yml", use_parent = TRUE)

########################################################################
#
# Logging setup
#
flog.threshold(INFO)
flog.appender(appender.file2(paste0("logs/myproject-",Sys.Date(),".log"), console=TRUE))
flog.info("Started...")
}
\keyword{dependencies}
\keyword{library}
\keyword{require}
