% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqatools.R
\name{checkDir}
\alias{checkDir}
\title{checkDir}
\usage{
checkDir(directory)
}
\arguments{
\item{directory}{the directory}
}
\description{
Checks if a directory exists if it does not , it will create it
}
\examples{
checkDir('C:\\\\TMP')
}
\keyword{directory}
