% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqatools.R
\name{debug.string}
\alias{debug.string}
\title{debug.string}
\usage{
debug.string(x, fileext, startcmd = "start")
}
\arguments{
\item{x}{the string to open}

\item{fileext}{the fileextension for the os-lookup}

\item{startcmd}{the start command, for osx/linux use open}
}
\description{
Debugs a large string and opens the associated handler defined by the operating system
}
\examples{
debug.string(x = "SELECT 1,2", fileext = '.sql')
}
\keyword{debug}
\keyword{large}
\keyword{string}
