% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqatools.R
\name{emailR}
\alias{emailR}
\title{emailR}
\usage{
emailR(to = "", subject = "", body = "", cc = "", bcc = "")
}
\arguments{
\item{to}{the to email address}

\item{body}{the body of the email}

\item{cc}{the carbon copy}

\item{bcc}{the blind carbon copy}
}
\description{
Uses the COM bridge to talk to the outlook application (COMCreate)
A simple R object interface to the email client.  Can be used for attachments or including tables (dataframes) as html and plain text.
}
\examples{
df<-data.frame(First_Name=c("John","Jane"),Last_Name=c("Doe","Foo"))
write.csv(df, gzfile('C:/email-attachment-test.csv.gz'),row.names=FALSE,na = "")
testEmail = emailR(to = "James_Banasiak@bcbsil.com", subject = "Test Email", body = "Test Body")
testEmail$addHTMLTable(description = "Description", df = df)
testEmail$addAttachment('C:/email-attachment-test.csv.gz')
testEmail$save()
testEmail$display()
testEmail$send()
}
\keyword{email}
