% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqatools.R
\name{fromTemplateString}
\alias{fromTemplateString}
\title{fromTemplateString}
\usage{
fromTemplateString(input, varlist)
}
\arguments{
\item{input}{the template file to read as a string}

\item{variableValueList}{a list of variables and values}
}
\description{
reads a string and replaces expressions with the format [variable] with a specified value
}
\examples{
fromTemplateString("Hello [VAR1]  this is [VAR2] or test [VAR3]",list('[VAR1]'=c("VAL1"), '[VAR2]'=paste(1,2,3,sep=','), '[VAR3]'="VAL3"))

}
\keyword{template}
