% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqatools.R
\name{isValidNPI}
\alias{isValidNPI}
\title{isValidNPI}
\usage{
isValidNPI(npi)
}
\arguments{
\item{npi}{the npi}
}
\description{
Computes a valid npi
}
\examples{
isValidNPI(1)
isValidNPI(1396860391)
}
\keyword{comments}
