% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqatools.R
\name{queue}
\alias{queue}
\title{queue}
\usage{
queue(statistics = FALSE)
}
\arguments{
\item{statistics}{record statistics}
}
\description{
Simple Queue Interface
}
\examples{
# create a new queue
myQueue<-queue(TRUE)
for(i in 1:100){
  myQueue$enqueue(i)
}
myQueue$summary()
myQueue$resetStats()
for(i in 1:50){
  myQueue$dequeue()
}
myQueue$summary()
for(i in 1:26){
  myQueue$enqueue(letters[i])
}
myQueue$summary()
}
\keyword{queue}
