% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqatools.R
\name{read.multisep}
\alias{read.multisep}
\title{read.multisep}
\usage{
read.multisep(filename, headers = NULL, sep = "|&~&|", n = -1L,
  skipNul = TRUE)
}
\arguments{
\item{filename}{the file}
}
\description{
reads a multi seperator file
}
\examples{
f<-'PAL_PFIN_TAX_ID.txt'
df1<-read.multisep(filename=f,n=10L,sep = '|&~&|',headers=paste('column',c(1:15)))

}
\keyword{comments}
