% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqatools.R
\name{removeExtension}
\alias{removeExtension}
\title{removeExtension}
\usage{
removeExtension(file)
}
\arguments{
\item{file}{the filename}
}
\description{
removes an extension from a file
}
\examples{
removeExtension("C:\\\\TEMP\\\\TESTFILE.123")
basename(removeExtension("C:\\\\TEMP\\\\TESTFILE.123"))
}
\keyword{extension}
