% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqatools.R
\name{removeSQLComments}
\alias{removeSQLComments}
\title{removeSQLComments}
\usage{
removeSQLComments(x)
}
\arguments{
\item{x}{the string to remove commments from}
}
\description{
Removes comments from sql
}
\examples{
removeSQLComments("
/***
multiline1
**/
             /* c++ style1 */
--single1
select 1 /*inside1*/,2
       /*
multiline2
*/


--single2
/* c++ style2 */

")
}
\keyword{comments}
