% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqatools.R
\name{sumdigits}
\alias{sumdigits}
\title{sumdigits}
\usage{
sumdigits(x)
}
\arguments{
\item{x}{the list of digits}
}
\description{
Sums individual digits - used in NPI checksum
}
\examples{
# where 1+2=3 and 2+7=9
sumdigits(c(12,27))
[1] 3 9
}
\keyword{comments}
