% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqatools.R
\name{teraDataOdbc}
\alias{teraDataOdbc}
\title{teraDataOdbc}
\usage{
teraDataOdbc(host = "edwp.hcscint.net", user = "", port = 1025,
  mode = "TERADATA", driver.name = "auto", driver.pattern = "Teradata")
}
\arguments{
\item{host}{the host to connect to defaults to PRODUCTION edwp.hcscint.net}

\item{port}{Teradata default 1025}

\item{mode}{Session ANSI or TERADATA mode}

\item{driver.name}{Drive name to use, with auto mode it will guess}

\item{driver.pattern}{The driver pattern to search for defaults to Teradata}
}
\description{
An R object interface to Teradata
This object can connect to teradata and perform tasks such as query and execute procedures.
}
\examples{
# initialize
td = teraDataOdbc(user='u365042', driver.name ='auto', driver.pattern="Teradata")
td$connect()

# simple
df<-td$query("SELECT 1,2")
print(df)

# fastexport a table (must use double backslashes in directory)
# the chop parameter is used to chop off the first two bytes using sed as an outmod
td$exportTables(schema = "ENTPR_EQA", tables = c("TMP_MEASURE_DIM"), directory = "C:\\\\TEMP", chop=TRUE)

# call stored proc
td$query("CALL GLOBAL_TOOLS.TD_COMPRESS('ENTPR_EQA','TMP_SUBMEASURE_DIM2','Y',RC,MSG,OUTPUT_MESSAGES,ALTER_STMT)")

# more complex
}
\keyword{teradata}
